ipaxi/requirements/test.txt
# Test dependencies go here.
-r base.txt


coverage==4.5.1
django-coverage-plugin==1.5.0
django-test-plus==1.0.22
django-debug-toolbar==1.9.1
factory_boy==2.10.0
factory_djoy==0.6.0

# pytest
pytest==3.5.0
pytest-django==3.1.2
pytest-sugar==0.9.1
pytest-xdist==1.22.2
pytest-cov==2.5.1

model-mommy==1.5.1
